Create menu package
Created menu.js. Apart from that, changed the mode to development
because i get to know that mode should be set to development during
making of the project since it takes less time to bundle the files. Also
installed MiniCssExtractPlugin and added it to webpack.config file as a
comment for now,latter when development will be completed then i will
change mode to production and uncomment this commented code.
MiniCssExtractPlugin helps to create a separate css file in the dist
folder, it helps to make js file small and makes css cashable in browser.